# Edit onnx parser to infer values in post order