@mbrookhart 2020-06-11
Yes, that's the idea, and why the hugging face model import speeds up so much. It seems like there is significant overhead each time we call MCJIT though, which is why BERT-Squad, with many ops but fewer infer_value calls, slows down.