@masahi 2020-06-11
ok I now understand why it makes importing some model slower. Previously there are fewer but heavy `infer_value` calls, now there are many but they are all single node compile, so LLVM overhead dominates.

I wish there were a "fast-path" compile option for llvm for use case like this. Turning on all optimization to compile a single node and get a compile time constant is clearly overkill. At least we can lower the opt level (always 3 in our backend) for const eval mode @tqchen