@masahi 2020-06-11
I think it's a very good idea, and we can probably mitigate the slow down by tweaking llvm.

Will leave it open for another day for others to comment 
@soiferj @siju-samuel @icemelon9 @kevinthesun